Use --ipv(4|6)-nameserver as argument consistently
This essentially reverts #445 and fixes it a bit different. First of,
the plural is confusing as multiple nameservers can't be passed using
the --ipv(4|6)-nameservers argument. Instead, specifying the argument
multiple times is the right way to go.

However, the API indeed uses `nameservers` as the key, and that is
why the argument change fixes the command in first place.

The Supervisor API is a bit inconsistent here: `address` is a list too,
but the key string is a singular. This makes the CLI a bit weird.

Use singular for the string array argument consistently for address and
nameserver as it feels more natural.

Also fix the nameserver argument for VLANs.