Increase timeout for operations requiring reboot to 90s (#556)
Users are sometimes confused when they try to reboot or switch slots and
the operation fails with the generic error, while it's been already
submitted and is performed shortly after. Allow more time for the
operation to finish gracefully, as terminating it while it's been posted
usually wouldn't cancel it anyway.