Add `--interactive` option for `ha auth reset`
Add option to perform interactive reset of user password. In that case,
when no username argument is provided, a list of existing users is shown
first. With username selected or given, prompt for a password is shown,
read with no echo to the terminal for privacy.

This can be especially useful when the username contains characters that
can't be entered in the OS console. While the framebuffer console still
may not show all the characters, it may give enough clues to select the
users for whom the password should be reset. See:
https://github.com/home-assistant/operating-system/issues/3879